requirements.txt
Django>=4.2,<5.0
openai>=1.0.0
python-dotenv>=1.0.0
djangorestframework>=3.14.0
drf-spectacular>=0.26.0
Pillow>=9.0.0
django-cors-headers>=4.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
redis>=4.0.0
celery>=5.3.0